requirements.txt
pillow
aiosqlite
aiohttp>=3.7.4,<4
qq-botpy
sqlalchemy
cryptography
requests
sqlalchemy
python-dotenv
pytest
pytest-asyncio
plotly
kaleido
pandas